QA: made as much sensible changes as possible after a     FindBugs run.
git-svn-id: https://svn.codehaus.org/izpack/izpack-src/trunk@850 7d736ef5-cfd4-0310-9c9a-b52d5c14b761

+++ src/lib/com/izforge/izpack/panels/CompilePanelAutomationHelper.java
@@ -168,7 +168,7 @@ public class CompilePanelAutomationHelper extends PanelAutomationHelper
    */
   public void progress(int val, String msg)
   {
-    float percentage = ((float)val)*100.0f/(float)this.job_max;
+    double percentage = ((double)val)*100.0d/(double)this.job_max;
 
     String percent = (new Integer ((int)percentage)).toString()+'%';
     String line = this.job_name + ": " + percent;
